+++ test/feature/dynamic_types/dds_types_tests/DynamicTypesStringsDDSTypesTests.cpp
@@ -53,7 +53,8 @@ TEST_F(DynamicTypesDDSTypesTest, DDSTypesTest_StringStruct)
 
     MemberDescriptor::_ref_type member_descriptor {traits<MemberDescriptor>::make_shared()};
     member_descriptor->name(var_string_name);
-    member_descriptor->type(DynamicTypeBuilderFactory::get_instance()->create_string_type(LENGTH_UNLIMITED)->build());
+    member_descriptor->type(DynamicTypeBuilderFactory::get_instance()->create_string_type(static_cast<uint32_t>(
+                LENGTH_UNLIMITED))->build());
     type_builder->add_member(member_descriptor);
     DynamicType::_ref_type struct_type {type_builder->build()};
 
@@ -86,7 +87,8 @@ TEST_F(DynamicTypesDDSTypesTest, DDSTypesTest_WStringStruct)
 
     MemberDescriptor::_ref_type member_descriptor {traits<MemberDescriptor>::make_shared()};
     member_descriptor->name(var_wstring_name);
-    member_descriptor->type(DynamicTypeBuilderFactory::get_instance()->create_wstring_type(LENGTH_UNLIMITED)->build());
+    member_descriptor->type(DynamicTypeBuilderFactory::get_instance()->create_wstring_type(static_cast<uint32_t>(
+                LENGTH_UNLIMITED))->build());
     type_builder->add_member(member_descriptor);
     DynamicType::_ref_type struct_type {type_builder->build()};
